made components folder
project page components are in a folder
