+++ app/page.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import AuthForm from "./auth-form"
 import { createRouteHandlerClient } from "@supabase/auth-helpers-nextjs"
-import { cookies, headers } from "next/headers"
+import { cookies } from "next/headers"
 import { redirect } from "next/navigation"
 
 const Home: React.FC = async () => {
