+++ package.json
@@ -24,6 +24,7 @@
     "openai": "^3.3.0",
     "postcss": "^8.4.24",
     "react": "^18.2.0",
+    "react-canvas-confetti": "^1.4.0",
     "react-dom": "^18.2.0",
     "react-xarrows": "^2.0.2",
     "tailwindcss": "^3.3.2"
